backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
pgvector>=0.2.4
sentence-transformers>=2.2.2
pymupdf>=1.23.0
python-docx>=1.1.0
groq>=0.4.0
httpx>=0.25.0
python-dotenv>=1.0.0
gunicorn>=21.0
